Moves business logic to source dir
We moved contract tests to a separate test folder inside test directory.
We separated application source code to a specific src folder to improve readability.
