+++ src/controllers/equipmentController.js
