+++ src/fetcher/canVariablesFetcher.js
