+++ src/fetcher/equipmentFetcher.js
