+++ src/parser/equipmentParser.js
