+++ src/schemas/equipmentSchema.js
