+++ test/consumer-contracts/contract_helper.js
