+++ test/consumer-contracts/contract_http_client.js
