+++ test/consumer-contracts/contracts/telemetry_api/equipment/equipment/get_list.js
