+++ test/consumer-contracts/contracts/telemetry_api/equipment/equipment/get_one_by_id.js
