+++ test/consumer-contracts/contracts/telemetry_api/equipment/trackingData/search_last_trackingPoint.js
@@ -1,14 +1,13 @@
 import { Contract, Joi } from 'consumer-contracts';
 import config from '../../../../contract_config';
-import trackingDataSearchHelper from '../../../../helper/trackingDataSearchHelper';
+import trackingDataSearchHelper from '../../../../../../src/helper/trackingDataSearchHelper';
 import helper from '../../../../contract_helper';
 
-const canVariables = 'ENGINE_HOURS';
 const EQUIPMENT_ID = '7d7ceeb8-f628-45a1-b96d-587b3979f8ef';
 const TRACKER_ID = '5818efa2-e095-430f-aef1-aaa2db38095e';
 
 module.exports = new Contract({
-  name: 'Search by canVariables type',
+  name: 'Search by last equipment trackingPoint and duty',
   consumer: 'Telemetry API',
   before: (done) => {
     helper.deleteEquipment(EQUIPMENT_ID)
@@ -32,7 +31,7 @@ module.exports = new Contract({
     headers: {
       authorization: config.ACCESS_TOKEN
     },
-    url: trackingDataSearchHelper.searchByCANVariablesUrl(canVariables, EQUIPMENT_ID)
+    url: trackingDataSearchHelper.searchLastDutyUrl(EQUIPMENT_ID)
   },
   response: {
     statusCode: 200,
@@ -56,10 +55,9 @@ module.exports = new Contract({
           timeOfReception: Joi.string(),
           externalId: Joi.string()
         }),
-        canVariables: Joi.array().items({
+        duties: Joi.array().items({
           id: Joi.string(),
-          canId: Joi.number(),
-          name: Joi.string()
+          status: Joi.string()
         })
       }),
       meta: Joi.object().keys({
@@ -67,12 +65,7 @@ module.exports = new Contract({
           equip_agg: Joi.array().items({
             key: Joi.string(),
             count: Joi.number(),
-            tp_latest_ag: Joi.array(),
-            spn_ag: Joi.array().items({
-              key: Joi.string().valid(canVariables),
-              count: Joi.number(),
-              spn_latest_ag: Joi.array()
-            })
+            tp_latest_ag: Joi.array()
           })
         })
       })
@@ -80,9 +73,9 @@ module.exports = new Contract({
   },
   after: (done) => {
     helper.deleteEquipment(EQUIPMENT_ID)
-    .then(() => {
-      return helper.deleteTracker(TRACKER_ID);
-    })
+     .then(() => {
+       return helper.deleteTracker(TRACKER_ID);
+     })
     .catch(() => {})
     .finally(done);
   }
