+++ test/controllers/equipment/options_equipment_spec.js
@@ -1,8 +1,8 @@
-import app from '../../../app';
-import CanVariablesFetcher from '../../../fetcher/canVariablesFetcher';
-import EquipmentFetcher from '../../../fetcher/equipmentFetcher';
+import app from '../../../src/app';
+import CanVariablesFetcher from '../../../src/fetcher/canVariablesFetcher';
+import EquipmentFetcher from '../../../src/fetcher/equipmentFetcher';
 import td from 'testdouble';
-import TrackingPointFetcher from '../../../fetcher/trackingPointFetcher';
+import TrackingPointFetcher from '../../../src/fetcher/trackingPointFetcher';
 
 describe('Asking for supported HTTP methods at (/equipment)', () => {
   let canVariablesFetcher;
